requirements.txt
# requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
requests==2.31.0
rtree==1.0.1
shapely==2.0.1
pyproj==3.5.0
geopandas==0.13.2
rasterio==1.3.8
networkx==3.1
python-dotenv==1.0.0
